reorganize functions used in matching users,add create weights function to make a weight from a list of ints
